+++ CHANGELOG.md
@@ -7,6 +7,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Upgraded stable Dart analysis SDK to `3.13.0`
  * Upgraded stable Flutter analysis SDK to `3.47.0`.
  * Upgraded dependencies.
+ * Note: started to mirror `neat_periodic_task` status data into SQL.
 
 ## `20260812t113900-all`
  * Bump runtimeVersion to `2026.08.12`.